Safari-609.1.11.1/LayoutTests/svg/animations/svgangle-animation-grad-to-deg-expected.txt

SVG 1.1 dynamic animation tests

Tests SVGAngle animation from grad to deg.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS marker.orientAngle.animVal.value is 0
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientType.animVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientAngle.animVal.value is 90
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientType.animVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientAngle.animVal.value is 180
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientType.animVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientAngle.animVal.value is 0
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientType.animVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS successfullyParsed is true

TEST COMPLETE
